// uvic-captions-player-block/src/types/jwplayer.ts
interface Environment {
  Browser: Browser;
  OS: OS;
  Features: Features;
}
// ...
type EventCallback<T> = (param: T) => void;

interface Region {
  x: 0; // x and y will always be 0 according to https://developer.jwplayer.com/jw-player/docs/javascript-api-reference/#jwplayergetsaferegion
  y: 0;
  width: number;
  height: number;
}

interface CaptionOptions {
  color: string;
  fontSize: number;
  fontFamily: string;
  fontOpacity: number;
  backgroundColor: string;
  backgroundOpacity: number;
  edgeStyle: string;
  windowColor: string;
  windowOpacity: number;
}

interface Level {
  bitrate: number;
  height: number;
  width: number;
  label: string;
}

interface QualityLevel {
  mode: "auto" | "manual";
  level: Level;
  reason: "auto" | "api" | "initial choice";
}

interface CastParam {
  available: boolean;
  active: boolean;
  deviceName: string;
  type: "cast";
}

interface WarningParam {
  code: number;
  message: string;
  sourceError: object | null;
  type: "warning";
}

interface EventParams {
  adClick: AdProgressParam;
  adCompanions: AdCompanionsParam;
  adComplete: AdProgressParam;
  adSkipped: AdProgressParam;
  adError: AdErrorParam;
  adRequest: AdRequestParam;
  adSchedule: AdScheduleParam;
  adStarted: AdStartedParam;
  adImpression: AdImpressionParam;
  adPlay: AdPlayParam;
  adPause: AdPlayParam;
  adTime: AdTimeParam;
  cast: CastParam;
  meta: MetadataParam;
  audioTracks: AudioTracksParam;
  audioTrackChanged: AudioTrackChangedParam;
  firstFrame: FirstFrameParam;
  buffer: BufferParam;
  bufferChange: BufferChangeParam;
  captionsChanged: CaptionsChangedParam;
  captionsList: CaptionsListParam;
  controls: ControlsParam;
  error: ErrorParam;
  fullscreen: FullscreenParam;
  idle: IdleParam;
  levelsChanged: LevelsChangedParam;
  mute: MuteParam;
  volume: VolumeParam;
  pause: PlayParam;
  play: PlayParam;
  playlist: PlaylistParam;
  playlistItem: PlaylistItemParam;
  ready: ReadyParam;
  resize: ResizeParam;
  visualQuality: VisualQualityParam;
  playbackRateChanged: PlaybackRateChangedParam;
  levels: LevelsParam;
  seek: SeekParam;
  setupError: ErrorParam;
  time: TimeParam;
  warning: WarningParam;
}

type NoParamEvent =
  | "adBlock"
  | "beforeComplete"
  | "complete"
  | "beforePlay"
  | "displayClick"
  | "playlistComplete"
  | "seeked"
  | "remove";

export interface JWPlayer {
  addButton(icon: string, label: string, handler: () => void, id: string): void;
  getAudioTracks(): any[];
  getBuffer(): number;
  getCaptionsList(): any[];
  getControls(): boolean;
  getCurrentAudioTrack(): number;
  getCurrentCaptions(): number;
  getCurrentQuality(): number;
  getDuration(): number;
  getHeight(): number;
  getFullscreen(): boolean;
  getMute(): boolean;
  getPlaylist(): any[];
  getPlaylistIndex(): number;
  getPlaylistItem(index?: number): any;
  getPosition(): number;
  getQualityLevels(): any[];
  getRenderingMode(): string;
  getSafeRegion(): Region;
  getState(): string;
  getVolume(): number;
  getContainer(): HTMLElement;
  getEnvironment(): Environment;
  getWidth(): number;
  getVisualQuality(): QualityLevel | undefined;
  getPlaybackRate(): number;
  load(playlist: any[] | string): void;
  on<TEvent extends keyof EventParams>(
    event: TEvent,
    callback: EventCallback<EventParams[TEvent]>
  ): void;
  on(event: NoParamEvent, callback: () => void): void;
  once<TEvent extends keyof EventParams>(
    event: TEvent,
    callback: EventCallback<EventParams[TEvent]>
  ): void;
  once(event: NoParamEvent, callback: () => void): void;
  off(event: keyof EventParams | NoParamEvent): void;
  off(event: NoParamEvent, callback: () => void): void;
  off<TEvent extends keyof EventParams>(
    event: TEvent,
    callback: EventCallback<EventParams[TEvent]>
  ): void;
  trigger<TEvent extends keyof EventParams>(
    event: TEvent,
    args: EventParams[TEvent]
  ): void;
  trigger(event: NoParamEvent): void;
  pause(state?: boolean): void;
  play(state?: boolean): void;
  playAd(tag: string): void;
  playlistItem(index: number): void;
  registerPlugin(
    id: string,
    target: string,
    jsPlugin: () => void,
    swfURL?: string
  ): void;
  remove(): void;
  removeButton(id: string): void;
  resize(width: number, height: number): void;
  seek(position: number): void;
  setControls(controls: boolean): void;
  setCurrentAudioTrack(index: number): void;
  setCurrentCaptions(index: number): void;
  setCurrentQuality(index: number): void;
  setPlaybackRate(rate: number): void;
  setFullscreen(state: boolean): void;
  setMute(state?: boolean): void;
  setup(options: any): JWPlayer;
  setVolume(volume: number): void;
  setCaptions(options: CaptionOptions): void;
  stop(): void;
}

export interface JWPlayerStatic {
  (query?: string | number | Element): JWPlayer;
  key: string;
  version: string;
}